Added more 22.GA functionality, such as `details_synchronization` and `children_synchronization`, and various fixes, now verified in the subdevices flow.

+++ examples/simple/Application.cpp
@@ -38,7 +38,7 @@ std::uint64_t generateRandomValue()
 {
     // Here we will create the random engine and distribution
     static auto engine =
-      std::mt19937(static_cast<std::uint64_t>(std::chrono::system_clock::now().time_since_epoch().count()));
+      std::mt19937(static_cast<std::uint32_t>(std::chrono::system_clock::now().time_since_epoch().count()));
     static auto distribution = std::uniform_real_distribution<>(-20, 80);
 
     // And generate a random value
